Fix hangs on on OSX
There were two separate (OSX-specific) issues causing this:

1. `mktemp` does not support the "suffix flag"
2. `echo` handles escaping different, and inserts a literal newline
   character into the python code.

As a result, we failed to start the python process. Since the process
would not read the FIFO, writing to the FIFO would hang. This commit
fixes issue (1) by attaching the suffix using string concatenation, and
fixes issue (2) by preferring printf, which is more portable.